Extract add remove rules tests from MiscTest class into separate test classes
